without loop, sum all squared values from 1 to n including n

+++ math/calculus/9-sum_total.py
@@ -3,5 +3,5 @@ def summation_i(n):
     if not isinstance(n, (int, float)):
         return None
     else:
-        total = n*(n + 1)(2*n + 1)/6
+        total = int(n*(n + 1)(2*n + 1)/6)
     return total
